feature: add like toggle api
